게시판 수정

+++ src/main/webapp/WEB-INF/views/board/board.jsp
@@ -18,51 +18,47 @@
                 <label class="form-label" for="writer">작성자 (작성일자 : <fmt:formatDate value="${postDto.createdDate}"
                                                                                    pattern="yyyy-MM-dd"
                                                                                    type="date"/>)</label>
-                <input id="writer" class="form-control" type="text" name="title" readonly value="${postDto.userId}">
+                <input id="writer" class="form-control" type="text" name="userId" readonly value="${postDto.userId}">
             </div>
         </div>
     </div>
     <div class="row mt-3">
         <div class="col-md-12">
-            <textarea readonly id="board_textarea" class="form-control mt-1" name="content" rows="20"
+            <textarea readonly class="form-control mt-1" name="content" rows="20"
                       cols="80">${postDto.content}</textarea>
         </div>
     </div>
     <input id="commentList" type="hidden">
+    <div class="input-group mb-3">
+        <input style="background-color: #e9eff5" id="inputComment" class="form-control" type="text"
+               placeholder="댓글을 입력하세요">
+        <button id="registerComment" onclick="registerComment()" type="button" class="btn btn-outline-primary">댓글 등록
+        </button>
+        <button style="display: none" id="modifyComment" type="button" class="btn btn-outline-primary">수정</button>
+        <button style="display: none" id="modifyCancel" type="button" class="btn btn-outline-danger">취소</button>
+    </div>
     <!-- 댓글 -->
     <div class="row mt-5">
         <div class="col-md-12" style="text-align: right">
-            <button onclick='location.href="<c:url
-                    value='/${postCategory}?page=${param.page}&option=${param.option}&keyword=${param.keyword}'/>"'
-                    class="btn btn-lg btn-outline-primary" type="button">목록
+            <button class="btn btn-lg btn-outline-primary" type="button" onclick='location.href="<c:url
+                    value='/${postCategory}?page=${param.page}&option=${param.option}&keyword=${param.keyword}'/>"'>목록
             </button>
-            <%--            <c:if test="${userId == postDto.userId}">--%>
-            <button id="modify" class="btn btn-lg btn-primary" type="button">수정</button>
-            <button id="remove" class="btn btn-lg btn-primary" type="button">삭제</button>
-            <%--            </c:if>--%>
+            <c:if test="${userId == postDto.userId}">
+                <button id="modify" class="btn btn-lg btn-primary" type="button">수정</button>
+                <button id="remove" class="btn btn-lg btn-primary" type="button">삭제</button>
+            </c:if>
         </div>
     </div>
 </div>
 
-<script src="https://cdn.ckeditor.com/ckeditor5/36.0.1/classic/ckeditor.js"></script>
 <script>
     // 글 수정
     $("#modify").on("click", function () {
         let isReadOnly = $("textarea[name=content]").attr("readonly");
         // 글 수정 시작
         if (isReadOnly === "readonly") {
-            // 에디터로 전환
-            ClassicEditor
-                .create(document.querySelector('#board_textarea'))
-                .then(editor => {
-                    editor.getData(${postDto.content})
-                    editor.ui.view.editable.element.style.minHeight = '500px';
-                })
-                .catch(error => {
-                    console.error(error);
-                });
-            $("#modifyPost").html("등록");
-            $("#writer").attr("readonly", true)
+            $("#modify").html("등록");
+            $("#remove").hide()
             $("input[name=title]").attr("readonly", false);
             $("textarea[name=content]").attr("readonly", false);
         }
@@ -75,7 +71,7 @@
             modData.content = $("textarea[name=content]").val()
             $.ajax({
                 type: "PATCH",
-                url: "/muscles/${postCategory}",
+                url: "/muscles/${postCategory}/" + postNo,
                 headers: {              // Http header
                     "Content-Type": "application/json",
                 },
@@ -114,6 +110,7 @@
     let postNo = ${postDto.postNo};
     // 댓글 불러오기
     loadComments()
+
     function loadComments() {
         $.ajax({
             type: "GET",            // HTTP method type(GET, POST) 형식이다.
@@ -122,6 +119,7 @@
                 "Content-Type": "application/json",
             },
             success: function (res) {
+                $(".comment-area").remove()
                 $("#commentList").after(toHtml(res))
             },
             error: function () {
@@ -129,20 +127,23 @@
             }
         })
     }
+
     // 댓글 생성
-    $("#createComment").on("click", function () {
+    function registerComment() {
         let userId = '${userId}';
         let content = $("#inputComment").val()
+        if (content === '') {
+            alert("댓글을 입력해주시기 바랍니다.")
+            return;
+        }
         $.ajax({
             type: "POST",            // HTTP method type(GET, POST) 형식이다.
-            url: "/muscles/comments/", // 컨트롤러에서 대기중인 URL 주소이다.
+            url: "/muscles/comments", // 컨트롤러에서 대기중인 URL 주소이다.
             headers: {              // Http header
                 "Content-Type": "application/json",
             },
             data: JSON.stringify({
-                userId: userId,
-                postNo: postNo,
-                content: content
+                userId: userId, postNo: postNo, content: content
             }),
             success: function () {
                 loadComments()
@@ -152,10 +153,12 @@
             }
         })
         $("#inputComment").val("")
-    })
+    }
+
     // 댓글 삭제
-    $("#commentList").on("click", ".delBtn", function () {
-        let commentNo = $(this).parent().attr('data-commentno')
+    $(document).on("click", ".delBtn", function () {
+        if (!confirm("정말로 삭제하시겠습니까?")) return;
+        let commentNo = $(this).parent().parent().attr('data-commentno')
         $.ajax({
             type: "DELETE",            // HTTP method type(GET, POST) 형식이다.
             url: "/muscles/comments/" + commentNo,
@@ -172,46 +175,58 @@
     })
 
     // 댓글 수정
-    $("#commentList").on("click", ".modBtn", function () {
-        let commentNo = $(this).parent().attr('data-commentNo')
-        $("#createComment").hide()
+    $(document).on("click", ".modBtn", function () {
+        $("#registerComment").hide()
         $("#modifyComment").show()
+        $("#modifyCancel").show()
+        document.body.scrollTop = document.body.scrollHeight;
+        let commentNo = $(this).parent().parent().attr('data-commentNo')
         let content = $(this).parent().children('.comment').text()
         $("#inputComment").val(content)
 
-        $("#modifyComment").on("click", function () {
+        $(document).on("click", "#modifyComment", function () {
             let content = $("#inputComment").val()
-            let data = {content: content}
-
+            let data = {commentNo: commentNo, content: content}
             $.ajax({
                 type: "PATCH",            // HTTP method type(GET, POST) 형식이다.
-                url: "/muscles/comments/" + commentNo,
+                url: "/muscles/comments",
                 headers: {              // Http header
                     "Content-Type": "application/json",
                 },
                 data: JSON.stringify(data),
                 success: function () {
+                    $("#registerComment").show()
+                    $("#modifyComment").hide()
+                    $("#modifyCancel").hide()
                     loadComments()
                 },
                 error: function () {
                     alert("AJAX 통신 실패")
                 }
             })
-            $("#createComment").show()
-            $("#modifyComment").hide()
             $("#inputComment").val("")
         })
     })
+    // 댓글 수정 취소
+    $("#modifyCancel").click(function () {
+        $("#registerComment").show()
+        $("#modifyComment").hide()
+        $("#modifyCancel").hide()
+        $("#inputComment").val("")
+    })
+
 
     let toHtml = function (comments) {
         let tmp = "";
         comments.forEach(function (comment) {
-            tmp += '<div class="row">'
+            tmp += '<div class="row comment-area">'
             tmp += '<div class="col-md-12" data-commentNo=' + comment.commentNo + ' data-postNo=' + comment.postNo + '>'
             tmp += '<div class="mb-0" style="border: 1px solid #e2e3e5">'
             tmp += '<span style="font-weight: bold" class="commenter">' + comment.userId + '</span>'
-            tmp += '<button style="float: right" type="button" class="delBtn btn btn-outline-primary">삭제</button>'
-            tmp += '<button style="float: right" type="button" class="modBtn btn btn-outline-primary">수정</button>'
+            if (comment.userId === '${userId}') {
+                tmp += '<button style="float: right" type="button" class="delBtn btn btn-outline-primary">삭제</button>'
+                tmp += '<button style="float: right" type="button" class="modBtn btn btn-outline-primary">수정</button>'
+            }
             tmp += ' <p style="font-style: italic" class="comment">' + comment.content + '</p>'
             if (comment.createdDate === comment.modDate)
                 tmp += ' <p style="font-weight: bold" class="commentDate">' + comment.createdDate
@@ -222,10 +237,6 @@
             tmp += '</div>'
             tmp += '</div>'
         })
-        tmp += '<div class="input-group mb-3">'
-        tmp += '<input style="background-color: #ced4da" id="inputComment" class="form-control" type="text" placeholder="댓글을 입력하세요">'
-        tmp += '<button id="createComment" type="button" class="btn btn-outline-primary">댓글 등록</button>'
-        tmp += '</div>'
         return tmp;
     }
 </script>
